created api package; created and shifted SyncTransactionChain related code to sync package; commented out SyncTransactionChainFrom function code in checks.go file;
